+++ js/thumbnails.js
